テーブルスキーマのプライマリキーが32bitになっている箇所を 64bitに拡張

+++ migrations/20210731022715-create-star.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER
+        type: Sequelize.BIGINT,
       },
       task_id: {
         type: Sequelize.BIGINT
